PHPStan Level 5
Summary: Mostly type related changes, but it actually found one bug

Test Plan: `./phpunit`

Reviewers: #hkccp_developers, mollekopf

Reviewed By: #hkccp_developers, mollekopf

Subscribers: #hkccp_developers

Differential Revision: https://git.kolab.org/D4622

+++ src/app/Console/Commands/Data/Import/LdifCommand.php
@@ -511,7 +511,7 @@ class LdifCommand extends Command
         $user->assignPackageAndWallet($this->packages['user'], $this->wallet ?: $user->wallets()->first());
 
         if (!empty($data->quota)) {
-            $quota = ceil($data->quota / 1024 / 1024) - $this->packages['quota'];
+            $quota = (int) (ceil($data->quota / 1024 / 1024) - $this->packages['quota']);
             if ($quota > 0) {
                 $user->assignSku($this->packages['storage'], $quota);
             }
